Remove obsolete resources and move ParametersResource to plans
